refactor: rename folder name feature -> features
