+++ src/features/post/model/hook.ts
