+++ src/features/post/model/index.ts
