+++ src/features/post/model/store.ts
