+++ src/features/post/type.ts
