+++ src/features/post/ui/PostTable.tsx
